bookstore-server-main/pyproject.toml
[tool.poetry]
name = "bookstore-server"
version = "0.1.0"
description = ""
authors = ["KatyaSavchuk <kateryna15savchuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"
drf-nested-routers = "^0.94.1"
django-cors-headers = "^4.6.0"
requests = "^2.32.3"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
